requirements.txt
python-docx
docx2txt
matplotlib
